Suppress age range on course info page
UCAS has 3 possible age ranges:
Primary (3-11/12)
Middle years (7 - 14 years)
Secondary (11+)

This age range doesn't fit all courses, but it's what's set in UCAS. When we show this to training providers they rightly point out that the age range is wrong and request changes.

We don't use this data, it doens't matter if it's correct. We can hide it for now to avoid confusion about age range and reduce support requests.