Code review: bring a property name check closer to property declaration
To generate the correct URL to the editor of an enrichment field,
we check if a field belongs to the "Contact" or the "About" section
by asking "does the property name equal any of the Contact properties"?

This moves this logic from the `.cshtml` to `OrganisationViewModelForContact`
where the tested properties are declared. This makes it more likely that
we keep this logic up to date.