Fix session timeout - set session length to 6 hours
There was an error in our cookie logic. We set the cookie lifetime
to equal the access_token lifetime. This meant that we lost access to the
cookie and the refresh token as soon as the access_token expired.
As a consequence, we never had a chance to refresh.

This decouples cookie lifetime from access_token lifetime, setting
the session length to 6 hours.